system page / server settings fully split
