+++ client_src/js/client-error.modal.js
@@ -1,11 +1,23 @@
 const octoFarmErrorModalElement = "#octofarmErrorModal";
 
+function returnDodoCodingMessage(options) {
+  return `
+     <div class="py-3">
+        Please report this error to <a href="https://github.com/octofarm/octofarm/issues">OctoFarm Issues</a>!
+     </div>
+      ${options.message}
+  `;
+}
+
 function returnErrorMessage(options) {
   return `
      <br>
-     ${options.type} ERROR (${options.statusCode}): 
+     ${options.type} ERROR (${options?.statusCode}): 
      <br>
      ${options.message}
+     <div class="py-3">
+        If this issue persists, please submit a bug report with the below information...
+     </div>
   `;
 }
 
@@ -24,16 +36,36 @@ function openErrorModal(options) {
   const apiErrorTitle = document.getElementById("apiErrorTitle");
   const apiErrorMessage = document.getElementById("apiErrorMessage");
   const apiDeveloperInfo = document.getElementById("apiDeveloperInfo");
-  apiErrorTitle.innerHTML = ` ${options.name}`;
+  apiErrorTitle.innerHTML = ` ${options?.name}`;
   apiErrorMessage.innerHTML = returnErrorMessage(options);
-  apiErrorMessage.className = `text-${options.color}`;
+  apiErrorMessage.className = `text-${options?.color}`;
   apiDeveloperInfo.innerHTML = returnModalDeveloperInfo(options);
   $(octoFarmErrorModalElement).modal("show");
 }
 
+function openDodoModal(options) {
+  //TODO: Duplicate, doesn't work when caught on page load comes back null :?
+  const apiErrorTitle = document.getElementById("apiErrorTitle");
+  const apiErrorMessage = document.getElementById("apiErrorMessage");
+  const apiDeveloperInfo = document.getElementById("apiDeveloperInfo");
+  apiErrorTitle.innerHTML = "Dodo Coding Error!";
+  apiErrorMessage.innerHTML = returnDodoCodingMessage(options);
+  apiErrorMessage.className = `text-${options?.color}`;
+  apiDeveloperInfo.innerHTML = returnModalDeveloperInfo(options.error);
+  $(octoFarmErrorModalElement).modal("show");
+}
+
+function handleEvent() {
+  if (!event.reason) {
+    openDodoModal(event);
+  } else {
+    openErrorModal(event.reason);
+  }
+}
+
 window.onunhandledrejection = function (event) {
-  openErrorModal(event.reason);
+  handleEvent(event);
 };
 window.onerror = function (event) {
-  openErrorModal(event.reason);
+  handleEvent(event);
 };
