+++ client_src/js/services/octofarm-client.service.js
@@ -50,10 +50,13 @@ axios.interceptors.response.use(
   }
 );
 
+// TODO: this could end up getting big, consider splitting it.
+// Would go by page, each page could get it's own extends class for pre-defined routes building on the CRUD actions available.
 export default class OctoFarmClient {
   static printerRoute = "/printers";
   static serverSettingsRoute = "/settings/server";
   static logsRoute = `${this.serverSettingsRoute}/logs`;
+  static updateSettingsRoute = `${this.serverSettingsRoute}/update`;
 
   static validatePath(pathname) {
     if (!pathname) {
@@ -96,7 +99,16 @@ export default class OctoFarmClient {
   }
 
   static async getServerSettings() {
-    return this.get("settings/server/get");
+    return this.get(`${this.serverSettingsRoute}/get`);
+  }
+
+  static async updateServerSettings(settingsObject) {
+    //TODO: should be patch not post
+    return this.post(this.updateSettingsRoute, settingsObject);
+  }
+
+  static async restartServer() {
+    return this.post(`${this.serverSettingsRoute}/restart`, {});
   }
 
   static async getCustomGcode() {
