+++ client_src/js/system/utils/bootbox.options.js
@@ -1,4 +1,8 @@
-import { setupOPTimelapseSettings } from "./server.actions";
+import {
+  setupOPTimelapseSettings,
+  restartOctoFarmServer,
+  updateOctoFarmCommand
+} from "../server.actions";
 
 const serverBootBoxOptions = {
   OP_TIMELAPSE_SETUP: {
@@ -21,6 +25,74 @@ const serverBootBoxOptions = {
         setupOPTimelapseSettings();
       }
     }
+  },
+  OF_SERVER_RESTART_REQUIRED: {
+    message: "Your settings changes require a restart, would you like to do this now?",
+    buttons: {
+      cancel: {
+        label: '<i class="fa fa-times"></i> Cancel'
+      },
+      confirm: {
+        label: '<i class="fa fa-check"></i> Confirm'
+      }
+    },
+    async callback(result) {
+      if (result) {
+        await restartOctoFarmServer();
+      }
+    }
+  },
+  OF_UPDATE_LOCAL_CHANGES: (updateOctoFarmBtn, message) => {
+    return {
+      title: '<span class="text-warning">Local file changes detected!</span>',
+      message: message,
+      buttons: {
+        cancel: {
+          className: "btn-danger",
+          label: '<i class="fa fa-times"></i> Cancel'
+        },
+        confirm: {
+          className: "btn-success",
+          label: '<i class="fa fa-check"></i> Override'
+        }
+      },
+      async callback(result) {
+        if (result) {
+          await updateOctoFarmCommand(true);
+        } else {
+          if (updateOctoFarmBtn) {
+            updateOctoFarmBtn.innerHTML = '<i class="fas fa-thumbs-up"></i> Update OctoFarm';
+            updateOctoFarmBtn.disabled = false;
+          }
+        }
+      }
+    };
+  },
+  OF_UPDATE_MISSING_DEPENDENCIES: (updateOctoFarmBtn, message) => {
+    return {
+      title: '<span class="text-warning">Missing dependencies detected!</span>',
+      message: message,
+      buttons: {
+        cancel: {
+          className: "btn-danger",
+          label: '<i class="fa fa-times"></i> Cancel'
+        },
+        confirm: {
+          className: "btn-success",
+          label: '<i class="fa fa-check"></i> Confirm'
+        }
+      },
+      callback: function (result) {
+        if (result) {
+          updateOctoFarmCommand(false, true);
+        } else {
+          if (updateOctoFarmBtn) {
+            updateOctoFarmBtn.innerHTML = '<i class="fas fa-thumbs-up"></i> Update OctoFarm';
+            updateOctoFarmBtn.disabled = false;
+          }
+        }
+      }
+    };
   }
 };
