+++ views/partials/System/systemClient.ejs
@@ -452,5 +452,5 @@
     </div>
 </div>
 <button type="button" class="btn btn-success btn-block mt-1"
-        id="saveSettings">Save
+        id="saveClientSettingsBtn">Save
 </button>
