+++ views/partials/octofarm-error.modal.ejs
@@ -9,9 +9,6 @@
             </div>
             <div class="modal-body text-center pt-0">
                 <div id="apiErrorMessage"></div>
-                <div class="py-3">
-                    If this issue persists, please submit a bug report with the below information...
-                </div>
                 <div id="apiDeveloperInfo"></div>
             </div>
         </div>
